Refuse to add an extraneous newline to old posts
The "bin/diff" utility warned me that the current formatting code, if
allowed to update my site, would add extraneous newlines to all of my
old HTML-source posts.  It turns out that this was because of an
embedded newline in the `legacy_intro` string constant, which this
commit hereby eliminates.