Add OpenConnect post HTML (whose source I lost?)
Why didn’t I check in the source code to this post when I wrote it?  And
where did it go?  After searching the repository history and `git stash`
for every laptop in the house that seems to have a checkout of my blog,
I’ve decided to just check in the HTML and go on with things.