SymPy eliminating a variable; new MathJax; Gentium
A couple of hours of yak shaving.

In a private email someone showed me how to use SymPy solve() on a
problem posed in a blog post, so I wanted to show the approach.  Alas,
my Notebook now renders more modern MathJax that my HTML was unprepared
for, so I upgraded to MathJax 3.  Alas, the math came out far too small,
and by CSS bisection I determined it was because of some problem with
the font “Andada” I was using.  I tried to turn the font off with a CSS
rule `mjx-container {font-family: serif}` which almost worked —
equations returned to normal size — but now superscripts were far too
big.  So: Andada is fired, and Gentium is now my font.  It’s a bit
spindly on a screen compared to Andada, but, it’s a pretty font and it
lets MathJax work without any fiddling.

Oh: also, updated how delimiters are handled, for the new MathJax code
emitted by SymPy.  At least it let me remove a special case.